Added base project and read model classes
AbstractReadModelProjector is a base read model projector for
our projectors. It launchs for any message the `any` method in
read model.

In read model, we check that an `applyMessageName` exists and
executes it in case.

AbstractReadModel needs to know how to init the database. That is a
responsability for the repository, so we need to returns an
SchemaManagerInterface.

SchemaManagerInterface should be implemented for any repository
injected in the read model.